rename lab1 to Lab1
