+++ src/lab1/lab1/Infrastructure/TypeRegistrar.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Spectre.Console.Cli;
 using System;
 
-namespace lab1.Infrastructure
+namespace Lab1.Infrastructure
 {
     public sealed class TypeRegistrar : ITypeRegistrar
     {
