+++ src/lab1/lab1/Models/Operations/BinaryOperation.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Serialization;
 
-namespace lab1
+namespace Lab1.Operations
 {
     [XmlInclude(typeof(Sum))]
     [XmlInclude(typeof(Div))]
