+++ src/lab1/lab1/Models/Operations/Mul.cs
@@ -4,19 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace lab1
+namespace Lab1.Operations
 {
     public class Mul : BinaryOperation
     {
         public override IntOperand Calculate(IntOperand first, IntOperand second) => new IntOperand(first * second);
 
-        public override bool Equals(object obj)
-        {
-            if (obj is not Mul)
-                return false;
-            else
-                return true;
-        }
+        public override bool Equals(object obj) => obj is not Mul;
 
         public override int GetHashCode() => nameof(Mul).GetHashCode();
