+++ src/lab1/lab1/Models/Operations/Sub.cs
@@ -4,19 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace lab1
+namespace Lab1.Operations
 {
     public class Sub : BinaryOperation
     {
         public override IntOperand Calculate(IntOperand first, IntOperand second) => new IntOperand(first - second);
 
-        public override bool Equals(object obj)
-        {
-            if (obj is not Sub)
-                return false;
-            else
-                return true;
-        }
+        public override bool Equals(object obj) => obj is not Sub;
 
         public override int GetHashCode() => nameof(Sub).GetHashCode();
