+++ src/lab1/lab1/Models/Operations/Sum.cs
@@ -5,20 +5,13 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Serialization;
 
-namespace lab1
+namespace Lab1.Operations
 {
     public class Sum: BinaryOperation
     {
         public override IntOperand Calculate(IntOperand first, IntOperand second) => new IntOperand(first + second);
 
-        public override bool Equals(object obj)
-        {
-            if (obj is not Sum)
-                return false;
-            else
-                return true;
-        }
-
+        public override bool Equals(object obj) => obj is not Sum;
         public override int GetHashCode() => nameof(Sum).GetHashCode();
 
         public override string ToString() => "sum";
