+++ src/lab1/lab1/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "lab1": {
       "commandName": "Project",
-      "commandLineArgs": "min"
+      "commandLineArgs": "add"
     }
   }
 }
