+++ src/lab1/lab1/Repositories/IBinaryOperationsRepository.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using Lab1.Operations;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace lab1.Repositories
+namespace Lab1.Repositories
 {
     interface IBinaryOperationsRepository
     {
